Revert "Revert "Extracted story validator in separate script""
This reverts commit 298360f62d0c5be2d13071d6e5a218ba20b4f55c.
